+++ bt/bt.cpp
@@ -14,8 +14,8 @@
 #include "app/config.h"
 #include "app/setup.h"
 #include "app/update_check.h"
-
 #include "win32/window.h"
+#include "app/rule_hit_log.h"
 
 #define OWN_WM_NOTIFY_ICON_MESSAGE WM_APP + 1
 
@@ -25,6 +25,7 @@ static const GUID NotifyIconGuid = {0x365f3f68, 0x6330, 0x4d4f, { 0xbe, 0xf2, 0x
 // globals.h
 alg::tracker t{APP_SHORT_NAME, APP_VERSION};
 lsignal::signal<void(const std::string&, const std::string&, const std::string&)> app_event;
+lsignal::signal<void(const bt::url_payload&, const bt::browser_match_result&)> open_on_match_event;
 
 using namespace std;
 
@@ -44,6 +45,13 @@ void execute(const string& data) {
         bool picker_down = bt::ui::is_picker_hotkey_down();
         bt::ui::open_method om = picker_down ? bt::ui::open_method::pick : bt::ui::open_method::configured;
         up.source_window_handle = (HWND)(DWORD)str::to_ulong(parts[1], 16);
+
+        win32::window win{up.source_window_handle};
+        up.window_title = win.get_text();
+
+        win32::process proc{win.get_pid()};
+        up.process_name = proc.get_name();
+
         bt::ui::url_open(up, om);
     } else {
         bt::ui::config();
@@ -99,6 +107,7 @@ int wmain(int argc, wchar_t* argv[], wchar_t* envp[]) {
 
     win32::popup_menu m{win32app.get_hwnd()};
     m.add("cfg", "Configure");
+    m.add("url_logger", "Open URL Log");
     m.add("$", "Buy me a Coffee!");
     m.add("contact", "Contact");
     m.add("?", "Help");
@@ -126,6 +135,8 @@ int wmain(int argc, wchar_t* argv[], wchar_t* envp[]) {
                 string id = m.id_from_loword_wparam(loword_wparam);
                 if(id == "cfg") {
                     bt::ui::config();
+                } else if(id == "url_logger") {
+                    bt::ui::url_logger();
                 } else if(id == "x") {
                     ::PostQuitMessage(0);
                 } else if(id == "$") {
@@ -167,12 +178,26 @@ int wmain(int argc, wchar_t* argv[], wchar_t* envp[]) {
             sni.display_notification("New Version", fmt::format("Version {} is now available!", arg1));
         } else if(name == "no_new_version") {
             sni.display_notification("Up to Date", "No new version available, you are up to date!");
-        } else if(name == "rule_open") {
-            if(bt::config::i.get_notify_on_rule_hit()) {
-                sni.display_notification("Link Opened", fmt::format("{}\n{}", arg1, arg2));
-            }
         }
     });
+
+    open_on_match_event.connect([&sni](const bt::url_payload& url, const bt::browser_match_result& bmr) {
+        if(bt::config::i.get_log_rule_hits()) {
+            bt::rule_hit_log::i.write(url, bmr);
+        }
+    });
+
+    // record basic stats
+    //{
+    //    t.track(map<string, string>
+    //    {
+    //        { "event", "open" },
+    //        {"browser", b->id},
+    //        {"profile_count", to_string(b->instances.size())},
+    //        {"rule_count", to_string(rules.size())},
+    //        {"arg", launch_arg},  // only record arg template here!
+    //    }, true);
+    //}
 
     // check for new versions
     string vn;